replace random_word with get_suffix(user)
